mod. poprawnosci spr istnienia pliku produktow

+++ src/main/java/pl/fourfun/readwriteproducts/CheckProductFile.java
@@ -8,11 +8,13 @@ import java.io.*;
 public class CheckProductFile {
 
     public static void checkingExistProductFile() throws IOException {
-        File productJsonFile = new File("Products.json");
-        productJsonFile.createNewFile();
-        FileOutputStream oFile = new FileOutputStream(productJsonFile, false);
 
-        JSONArray jsonArrayToFile = new JSONArray();
+        File productJsonFile = new File("Products.json");
+        if (!productJsonFile.exists()) {
+            productJsonFile.createNewFile();
+            FileOutputStream oFile = new FileOutputStream(productJsonFile, false);
+
+            JSONArray jsonArrayToFile = new JSONArray();
             JSONObject jsonObjectToFile = new JSONObject();
             jsonObjectToFile.put("productList", jsonArrayToFile);
 
@@ -26,25 +28,30 @@ public class CheckProductFile {
                 e.printStackTrace();
             }
         }
+    }
+
 
 
     public static void checkingExistUserFile() throws IOException {
+
         File userProductList = new File("UserProductList.json");
-        userProductList.createNewFile();
-        FileOutputStream oFile = new FileOutputStream(userProductList, false);
+        if (!userProductList.exists()) {
+            userProductList.createNewFile();
+            FileOutputStream oFile = new FileOutputStream(userProductList, false);
 
-        JSONArray jsonArrayToFile = new JSONArray();
-        JSONObject jsonObjectToFile = new JSONObject();
-        jsonObjectToFile.put("productList", jsonArrayToFile);
+            JSONArray jsonArrayToFile = new JSONArray();
+            JSONObject jsonObjectToFile = new JSONObject();
+            jsonObjectToFile.put("productList", jsonArrayToFile);
 
-        //Write JSON file
-        try (FileWriter file = new FileWriter("UserProductList.json")) {
+            //Write JSON file
+            try (FileWriter file = new FileWriter("UserProductList.json")) {
 
-            file.write(jsonObjectToFile.toJSONString());
-            file.flush();
+                file.write(jsonObjectToFile.toJSONString());
+                file.flush();
 
-        } catch (IOException e) {
-            e.printStackTrace();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
     }
 }
